+++ src/reactive/index.ts
@@ -1,6 +1,6 @@
 // TODO: "effect stack", knowing the stack of effects, to know when we are in a re-evaluation the stack like if we just entered from first effect execution of all effects in lineage
 
-export { lift, type ScanResult, scan } from './buffer'
+export { attend, lift, type ScanResult, scan } from './buffer'
 export { getState, touched, touched1 } from './change'
 export { deepWatch } from './deep-watch'
 export { cleanedBy, derived, effectAggregator, getActiveEffect } from './effect-context'
@@ -13,6 +13,7 @@ export {
 	getActivationLog,
 	onEffectThrow,
 	onEffectTrigger,
+	reset,
 	root,
 	untracked,
 } from './effects'